split some "jest-snapshot" utility functions to its own package "@jest/snapshot-utils"
