+++ packages/jest-snapshot-utils/package.json
@@ -1,0 +1,37 @@
+{
+  "name": "@jest/snapshot-utils",
+  "version": "30.0.0-alpha.4",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/jestjs/jest.git",
+    "directory": "packages/jest-snapshot-utils"
+  },
+  "license": "MIT",
+  "main": "./build/index.js",
+  "types": "./build/index.d.ts",
+  "exports": {
+    ".": {
+      "types": "./build/index.d.ts",
+      "require": "./build/index.js",
+      "import": "./build/index.mjs",
+      "default": "./build/index.js"
+    },
+    "./package.json": "./package.json"
+  },
+  "dependencies": {
+    "@jest/types": "workspace:*",
+    "chalk": "^4.0.0",
+    "graceful-fs": "^4.2.9",
+    "natural-compare": "^1.4.0"
+  },
+  "devDependencies": {
+    "@types/graceful-fs": "^4.1.3",
+    "@types/natural-compare": "^1.4.0"
+  },
+  "engines": {
+    "node": "^16.10.0 || ^18.12.0 || >=20.0.0"
+  },
+  "publishConfig": {
+    "access": "public"
+  }
+}
