+++ packages/jest-snapshot-utils/src/types.ts
@@ -1,0 +1,1 @@
+export type SnapshotData = Record<string, string>;
