+++ packages/jest-snapshot/src/State.ts
@@ -6,21 +6,23 @@
  */
 
 import * as fs from 'graceful-fs';
+import {
+  type SnapshotData,
+  getSnapshotData,
+  keyToTestName,
+  saveSnapshotFile,
+  testNameToKey,
+} from '@jest/snapshot-utils';
 import type {Config} from '@jest/types';
 import {getStackTraceLines, getTopFrame} from 'jest-message-util';
 import {saveInlineSnapshots} from './InlineSnapshots';
-import type {InlineSnapshot, SnapshotData, SnapshotFormat} from './types';
+import type {InlineSnapshot, SnapshotFormat} from './types';
 import {
   addExtraLineBreaks,
-  getSnapshotData,
-  keyToTestName,
   removeExtraLineBreaks,
   removeLinesBeforeExternalMatcherTrap,
-  saveSnapshotFile,
   serialize,
-  testNameToKey,
 } from './utils';
-
 export type SnapshotStateOptions = {
   readonly updateSnapshot: Config.SnapshotUpdateState;
   readonly prettierPath?: string | null;
